move machine learning into pro
