+++ operators/src/pro/engine/execution_context.rs
@@ -1,13 +1,12 @@
 use crate::engine::{
-    CreateSpan, ExecutionContext, InitializedPlotOperator, InitializedRasterOperator,
-    InitializedVectorOperator, MetaData, MetaDataProvider, MockExecutionContext, ResultDescriptor,
-    WorkflowOperatorPath,
+    CreateSpan, ExecutionContext, ExecutionContextExtensions, InitializedPlotOperator,
+    InitializedRasterOperator, InitializedVectorOperator, MetaData, MetaDataProvider,
+    MockExecutionContext, ResultDescriptor, WorkflowOperatorPath,
 };
 use crate::pro::meta::wrapper::InitializedOperatorWrapper;
 use crate::util::Result;
 use async_trait::async_trait;
 use geoengine_datatypes::dataset::{DataId, NamedData};
-use geoengine_datatypes::ml_model::MlModelId;
 use geoengine_datatypes::raster::TilingSpecification;
 use geoengine_datatypes::util::test::TestDefault;
 use rayon::ThreadPool;
@@ -63,22 +62,12 @@ impl ExecutionContext for StatisticsWrappingMockExecutionContext {
         op
     }
 
-    async fn load_ml_model(&self, model_id: MlModelId) -> Result<String> {
-        self.inner.load_ml_model(model_id).await
-    }
-
-    async fn store_ml_model_in_db(
-        &mut self,
-        model_id: MlModelId,
-        ml_model_str: String,
-    ) -> Result<()> {
-        self.inner
-            .store_ml_model_in_db(model_id, ml_model_str)
-            .await
-    }
-
     async fn resolve_named_data(&self, data: &NamedData) -> Result<DataId> {
         self.inner.resolve_named_data(data).await
+    }
+
+    fn extensions(&self) -> &ExecutionContextExtensions {
+        self.inner.extensions()
     }
 }
