+++ services/src/handlers/mod.rs
@@ -15,8 +15,6 @@ pub mod datasets;
 pub mod ebv;
 pub mod layers;
 
-#[cfg(feature = "xgboost")]
-pub mod model_training;
 pub mod plots;
 pub mod projects;
 pub mod session;
