+++ services/src/lib.rs
@@ -35,7 +35,6 @@ pub mod datasets;
 pub mod error;
 pub mod handlers;
 pub mod layers;
-pub mod machine_learning;
 pub mod ogc;
 pub mod projects;
 #[cfg(not(feature = "pro"))]
