+++ services/src/pro/handlers.rs
@@ -1,3 +1,5 @@
 pub mod datasets;
+#[cfg(feature = "xgboost")]
+pub mod model_training;
 pub mod permissions;
 pub mod users;
